Undo client supplied thread pools, revert some of the changes in PR that stopped catching Throwable
