+++ core/src/main/java/org/apache/accumulo/core/clientImpl/TabletServerBatchReaderIterator.java
@@ -386,15 +386,12 @@ public class TabletServerBatchReaderIterator implements Iterator<Entry<Key,Value
           fatalException = new TableDeletedException(tableId.canonical());
       } catch (SampleNotPresentException e) {
         fatalException = e;
-      } catch (Exception t) {
+      } catch (Throwable t) {
         if (queryThreadPool.isShutdown())
           log.debug("Caught exception, but queryThreadPool is shutdown", t);
         else
           log.warn("Caught exception, but queryThreadPool is not shutdown", t);
         fatalException = t;
-      } catch (Throwable t) {
-        fatalException = t;
-        throw t; // let uncaught exception handler deal with the Error
       } finally {
         semaphore.release();
         Thread.currentThread().setName(threadName);
@@ -411,7 +408,7 @@ public class TabletServerBatchReaderIterator implements Iterator<Entry<Key,Value
               e.setTableInfo(getTableInfo());
               log.debug("{}", e.getMessage(), e);
               fatalException = e;
-            } catch (Exception t) {
+            } catch (Throwable t) {
               log.debug("{}", t.getMessage(), t);
               fatalException = t;
             }
